+++ test/nixpkgs.nix
@@ -29,6 +29,10 @@ let
         # This test that a stack.yaml with no local packages defined is still
         # able to be used by stacklock2nix, and produces a reasonable package set.
         no-local-packages = final.callPackage ./test-no-local-packages {};
+
+        # This tests that leaving out the `packages` top-level key from `stack.yaml` will
+        # default to just using a single package in the same directory as the `stack.yaml` file.
+        default-local-package = final.callPackage ./test-default-local-package {};
       };
 
       # A list of all stacklock2nix tests.  This makes it easy to build all